Add 'has_cover' column for album (read detailed comment about this)
It greatly improves performance. Compare:
Without 'has_cover' property:
1. browser requests 'cover/thumb/777'
2. nginx searches 'cover/thumb/777.jpg' and 'cover/thumb/777.png'
3. if nothing found, nginx sends that request to tulip
4. tulip gets album and artist records (two or more SQL requests)
   because it needs full directory path
5. then it looks for cover.jpg/cover.png in album directory
6. if nothing found, it tries to extract cover from first .mp3 file
7. if still nothing found, tulip redirects to 'cover/thumb/0.png'
   which is 'coverart placeholder'

SEVEN STEPS! 6th step is the slowest one (I think)
And this is repeated over and over again for EACH ALBUM ON THE PAGE.
If my search returned 100 albums, that will be:
* 100 HTTP requests (from browser to nginx to tulip)
* 200 (or 300) unwated SQL requests
* 500 file lookups (cover.jpg/cover.png/folder.jpg/etc)
* 100 readings of .mp3 files to find cover in it!!!
That's pretty bad.

Now, with that new 'has_cover' property we can set source of image
to 0.png right away! Nginx will show placeholder or actual cover art
without redirecting requests to tulip. So, now we have:
* 1 HTTP request (from browser to nginx; without redirecting to tulip)
* 0 SQL requests
* 1 file lookup
* 0 readings of .mp3 files

Which is great improvement =)
All we need is to set has_cover to '1' each time when we add cover to album.
And that's pretty easy =)