v4.0: HUGE Refactoring (abolishing file's info in abyss folders) ++
We had a lot of complicated (spaghetti code) in 'folder' model.
All ratings info was stored in Folder.files::jsonb property.
And these information were partially transferred to 'records'
table when we are linking this folder to some particular release.
This was bad solution to the problem because:
* we are starting to have inconsistent data (rating in abyss can
  be different from rating in our library because synchronization
  worked only in one-way)
* to see how we rated other tracks (which are not in our library),
  we have to go to abyss folder and see. Same if we want to review
  ratings (and maybe add some more tracks for this release). It
  was cumbersome

So now, we are saving information about ALL records from abyss in
our 'records' table. That means that we'll have access to unrated
(and rated negatively) tracks anytime if we still have access to
these particular folder in abyss on our disk. And when we've lost
access to our abyss folder (due to disk corruption, etc), we still
have our rating's info saved (so we can see if we rated rest of
the tracks negatively or just haven't decided yet)
This also allows us to get rid of a lot of duplicate logic (eg:
assign rating methods)

Also I've deleted all deprecated chunks of code which were left
from previous (let me say) 'major version' of tulip (when we had
bunch of static pages without vue.js)

So this update can be officially called Tulip v.4.0

PS: Work is still in progress and full transition will be completed
if future commits