Disable 'keep-alive' for pages
As we need to keep the data updated (eg. when release cover is changed,
'keep-alive' prevented 'ReleaseComponent' to be updated, etc)

Also we had too much cached components, and we don't need them
to be cached 90% of the time (it's either useless, or causing wrong
behavior as described above)

PS: Page transitions got so much faster! Going back and forth to the
artists index/ releases index feels like 10x faster